+++ CountingSort.h
@@ -9,7 +9,7 @@ class CountingSort :
 public:
 	CountingSort();
 	~CountingSort();
-	void sort(std::vector<int>&) override;
+	void sort(std::vector<long long int>&) override;
 	void countingSort(int, int);
 	void print() override;
 	void printRunningTime(std::string&) override;
